Keep relation direction when updating content.
_This scratches my personal itch, where I want to nudge my Bolt-powered blog slightly more towards a knowledge-base thingy, where I can keep track of posts linking to other posts. To do this properly, I want to distinguish between incoming and outgoing links for a post._

Currently, a Content can have Relations To and From itself. In the edit interface, they're all thrown onto one pile, and when saving such an entry, all the relations will be, from that point on, From that content (even if they were in fact To it, before). This change fixes that, by more carefully storing and removing Relations. 

- Only remove existing Relations if they are actually no longer used
- Only add new Relations if they weren't already present

In addition:

- Don't return anything (and change the method signature), since the output was never actually used.