Implement switchable bidirectionality for related content.
This switches the boolean (which defaulted to `true` and never actually did anything) to a string, either
"to", "from" or "both" (the latter being the default). Choosing "to" or "from" limits the related content
to content that either relates _to_ us, or _from_ us. Choosing "both" gets you either.

For backwards compatibility purposes, you can also supply a boolean, where `true` will map to "both" and
`false` will map to "to".