Make both ECS and PHPStan happy.
The "fix" that ecs proposes means making the docblock... incomplete? It simply
removes a couple of the nullable parameters. If I list those as `?int $limit`,
for example, ecs is happy, but PHPStan complains. The only way to shut both
of them up is... not having a docblock in the first place. Which is how it's done
for other methods in this class, so I guess :man_shrugging: .