Fix bolt:setup by running doctrine scripts in new context
Doctrine caches the found migrations (which is logical), but that breaks
when you generate a new migration and then try to run it in the same
Application context. This was not shown on the console due to the
usage of the NullOutput. By migrating to the existing Script helper
we resolve the issue as each invocation now receives it's own
Application context without cached migrations.